[ADD] HealthController, used for checking the state of the pod inside k8s [ADD] DeleteControllerHanlder, handles all API request for the DeleteController [CHANGE] DependantResourceHandler, handles the deletion requests of all resources in parallel [ADD] Extensions and utils, reduce the boiler plate code within the application [CHANGE] Adjusted the MarkSessionClient and MarkSessionHandler according to the newest changes in the marking-svc [REMOVE] ErrorService [ADD] ErrorHandlingMiddleware, takes care of the exceptions within the application [ADD] Hangfire dependency, used to run the long running deletion process

+++ mars-deletion-svc/ResourceTypes/SimPlan/SimPlanClient.cs
@@ -4,6 +4,7 @@ using mars_deletion_svc.Exceptions;
 using mars_deletion_svc.MarkingService.Models;
 using mars_deletion_svc.ResourceTypes.SimPlan.Interfaces;
 using mars_deletion_svc.Services.Inerfaces;
+using mars_deletion_svc.Utils;
 
 namespace mars_deletion_svc.ResourceTypes.SimPlan
 {
@@ -21,20 +22,49 @@ namespace mars_deletion_svc.ResourceTypes.SimPlan
             _loggerService = loggerService;
         }
 
-        public async Task DeleteResource(DependantResourceModel dependantResourceModel)
+        public async Task DeleteResource(
+            DependantResourceModel dependantResourceModel,
+            string projectId
+        )
         {
-            var response = await _httpService.DeleteAsync(
-                $"http://sim-runner-svc/simplan?simPlanId={dependantResourceModel.ResourceId}"
-            );
-
-            if (response.StatusCode != HttpStatusCode.Accepted)
+            if (await DoesSimPlanExist(dependantResourceModel.ResourceId, projectId))
             {
-                throw new FailedToDeleteResourceException(
-                    $"Failed to delete {dependantResourceModel} from sim-runner-svc!"
+                var response = await _httpService.DeleteAsync(
+                    $"http://sim-runner-svc/simplan?simPlanId={dependantResourceModel.ResourceId}"
+                );
+
+                response.ThrowExceptionIfNotSuccessfulResponseOrNot404Response(
+                    new FailedToDeleteResourceException(
+                        $"Failed to delete {dependantResourceModel} from sim-runner-svc! The response status code is {response.StatusCode}"
+                    )
                 );
             }
 
             _loggerService.LogDeleteEvent(dependantResourceModel.ToString());
+        }
+
+        private async Task<bool> DoesSimPlanExist(
+            string simPlanId,
+            string projectId
+        )
+        {
+            var response = await _httpService.GetAsync(
+                $"http://sim-runner-svc/simplan?id={simPlanId}&projectid={projectId}"
+            );
+
+            if (response.IsSuccessStatusCode)
+            {
+                return response.StatusCode != HttpStatusCode.NoContent;
+            }
+
+            if (response.StatusCode == HttpStatusCode.NotFound)
+            {
+                return false;
+            }
+
+            throw new FailedToGetResourceException(
+                $"Failed to get simPlan with id: {simPlanId}, projectId: {projectId} from sim-runner-svc!"
+            );
         }
     }
 }
