+++ mars-deletion-svc/ResourceTypes/SimRun/SimRunClient.cs
@@ -4,6 +4,7 @@ using mars_deletion_svc.Exceptions;
 using mars_deletion_svc.MarkingService.Models;
 using mars_deletion_svc.ResourceTypes.SimRun.Interfaces;
 using mars_deletion_svc.Services.Inerfaces;
+using mars_deletion_svc.Utils;
 
 namespace mars_deletion_svc.ResourceTypes.SimRun
 {
@@ -21,27 +22,49 @@ namespace mars_deletion_svc.ResourceTypes.SimRun
             _loggerService = loggerService;
         }
 
-        public async Task DeleteResource(DependantResourceModel dependantResourceModel)
+        public async Task DeleteResource(
+            DependantResourceModel dependantResourceModel,
+            string projectId
+        )
         {
-            // TODO: It seems that the sim-runner-svc is not able to delete a simRun!!!
-            var response = await _httpService.DeleteAsync(
-                $"http://sim-runner-svc/simrun?simRunId={dependantResourceModel.ResourceId}"
-            );
-
-            if (response.StatusCode == HttpStatusCode.NotFound)
+            if (await DoesSimRunExist(dependantResourceModel.ResourceId, projectId))
             {
-                _loggerService.LogSkipEvent(dependantResourceModel.ToString());
-                return;
-            }
+                var response = await _httpService.DeleteAsync(
+                    $"http://sim-runner-svc/simrun?simRunId={dependantResourceModel.ResourceId}"
+                );
 
-            if (response.StatusCode != HttpStatusCode.OK)
-            {
-                throw new FailedToDeleteResourceException(
-                    $"Failed to delete {dependantResourceModel} from sim-runner-svc!"
+                response.ThrowExceptionIfNotSuccessfulResponseOrNot404Response(
+                    new FailedToDeleteResourceException(
+                        $"Failed to delete {dependantResourceModel} from sim-runner-svc! The response status code is {response.StatusCode}"
+                    )
                 );
             }
 
             _loggerService.LogDeleteEvent(dependantResourceModel.ToString());
+        }
+
+        private async Task<bool> DoesSimRunExist(
+            string simRunId,
+            string projectId
+        )
+        {
+            var response = await _httpService.GetAsync(
+                $"http://sim-runner-svc/simrun?simRunId={simRunId}&projectid={projectId}"
+            );
+
+            if (response.IsSuccessStatusCode)
+            {
+                return response.StatusCode != HttpStatusCode.NoContent;
+            }
+
+            if (response.StatusCode == HttpStatusCode.NotFound)
+            {
+                return false;
+            }
+
+            throw new FailedToGetResourceException(
+                $"Failed to get simRun with id: {simRunId}, projectId: {projectId} from sim-runner-svc!"
+            );
         }
     }
 }
